+++ src/react/components/tag/XUITag.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import cn from 'classnames';
 
 import '../helpers/xuiGlobalChecks';
-import { baseClass, variants } from './private/constants';
+import { baseClass, variants, sizes } from './private/constants';
 
 import '../../../sass/7-components/_typography.tags.scss';
 
@@ -12,12 +12,14 @@ const XUITag = ({
 	variant,
 	qaHook,
 	children,
+	size,
 }) => (
 	<span
 		className={cn(
 			baseClass,
 			className,
 			variants[variant],
+			sizes[size],
 		)}
 		role="status"
 		data-automationid={qaHook}
@@ -30,12 +32,15 @@ XUITag.propTypes = {
 	children: PropTypes.node,
 	qaHook: PropTypes.string,
 	className: PropTypes.string,
-	/* Type of tag to render */
+	/** Variant of tag to render */
 	variant: PropTypes.oneOf(Object.keys(variants)),
+	/** Size of tag to render */
+	size: PropTypes.oneOf(Object.keys(sizes)),
 };
 
 XUITag.defaultProps = {
 	variant: 'standard',
+	size: 'standard',
 };
 
 export { XUITag as default, variants };